refactor(texts): services file dialog delete all notes (#153)
* refactor(texts): service delete all texts

* style(texts): code style

* refactor(texts): rename folder

* refactor(texts): absolute import

* refactor(texts): export service

* refactor(texts): export service
