+++ src/modules/notes/Note/NoteView.tsx
@@ -1,16 +1,26 @@
-import CardNote from "./components/CardNote";
-import DialogDeleteAllNotes from "./components/Dialogs/DialogDeleteAllNotes";
-import DialogDeleteOneNote from "./components/Dialogs/DialogDeleteOneNote";
-import HeaderNote from "./components/HeaderNote";
+import Loading from "shared/components/Loading";
+import { ContextProviderNote } from "modules/notes/Note/Context";
+import CardNote from "modules/notes/Note/components/CardNote";
+import DialogDeleteAllNotes from "modules/notes/Note/components/Dialogs/DialogDeleteAllNotes";
+import DialogDeleteOneNote from "modules/notes/Note/components/Dialogs/DialogDeleteOneNote";
+import HeaderNote from "modules/notes/Note/components/HeaderNote";
+import DoodleMessage from "shared/components/DoodleMessage";
+import { INoteView } from "modules/notes/Note/types";
 
-const NoteView = () => {
+const NoteView: React.FC<INoteView> = ({ data, isFetching }) => {
 	return (
-		<>
-			<HeaderNote />
-			<CardNote />
+		<ContextProviderNote>
+			<Loading isLoading={isFetching} backdrop />
+			<HeaderNote count={data?.count} />
+			<DoodleMessage
+				title="You have not created any notes"
+				message="Create a note 😊"
+				show={data?.count === 0}
+			/>
+			<CardNote {...{ data, isFetching }} />
 			<DialogDeleteOneNote />
 			<DialogDeleteAllNotes />
-		</>
+		</ContextProviderNote>
 	);
 };
