+++ src/modules/notes/Note/components/CardNote/CardNote.tsx
@@ -1,40 +1,120 @@
-import { useTheme } from "@mui/material";
-import { useQuery } from "@tanstack/react-query";
-import { useContextNote } from "modules/notes/Note/Context";
-import CardNoteView from "modules/notes/Note/components/CardNote/CardNoteView";
-import { fetchGetNotesService } from "modules/notes/Note/components/CardNote/services";
-import { TypeCount } from "modules/notes/Note/components/CardNote/types";
-import { NOTE_QUERY_KEY } from "modules/notes/Note/constants";
+import { Grid, IconButton, Typography, useTheme } from "@mui/material";
+import {
+	Card,
+	CardAction,
+	CardContent,
+	CardDot,
+	CardHeader,
+	CardNoteContainer,
+} from "modules/notes/Note/components/CardNote/style";
 import { useDialogNote } from "modules/notes/Note/shared/hook/useDialogNote";
-import React, { memo, useEffect } from "react";
-import { useParams } from "react-router-dom";
+import React, { memo } from "react";
+import { FiEye, FiTrash } from "react-icons/fi";
+import { Link, useParams } from "react-router-dom";
+import TodoCount from "shared/components/TodoCount";
+import { ICON_SIZE } from "shared/constants";
+import { createURLQueryParams } from "shared/util/createURLQueryParams";
+import dateFormat from "shared/util/dateFormat";
+import { ICardNotes } from "./types";
 
-const CardNote: React.FC = () => {
+const CardNote: React.FC<ICardNotes> = ({ data }) => {
 	const { board_id } = useParams();
-	const { setCountNotes } = useContextNote();
 	const { palette } = useTheme();
+
 	const { openDialogDeleteOneNote } = useDialogNote();
 
-	const queryKey = [NOTE_QUERY_KEY.FETCH_GET_NOTES, { variable: board_id }];
-	const queryFn = () => fetchGetNotesService(board_id);
-
-	const onSuccess = ({ count }: TypeCount) => setCountNotes(count);
-	const { data, isFetching } = useQuery(queryKey, queryFn, { onSuccess });
-
-	useEffect(() => {
-		data && setCountNotes(data.count);
-	}, [board_id]);
-
 	return (
-		<CardNoteView
-			{...{
-				board_id,
-				palette,
-				data,
-				isFetching,
-				openDialogDeleteOneNote,
-			}}
-		/>
+		<Grid container spacing={2}>
+			{data &&
+				data.list_notes?.map(
+					({
+						note_id,
+						title_note,
+						observation,
+						color_note,
+						created_at,
+						todos,
+					}) => {
+						const createdAt = dateFormat(created_at);
+						const linkToNoteEdit = createURLQueryParams(
+							"/note/edit",
+							{
+								note_id,
+								board_id,
+							}
+						);
+						return (
+							<Grid item xl={2} md={3} xs={12} key={note_id}>
+								<CardNoteContainer>
+									<Card sx={{ height: 130 }}>
+										<CardContent>
+											<CardHeader>
+												<CardDot color={color_note} />
+												<Typography
+													variant="body1"
+													fontWeight={800}>
+													{title_note}
+												</Typography>
+											</CardHeader>
+
+											<Typography
+												variant="body2"
+												fontSize={13}
+												color="GrayText"
+												sx={{ mt: 1 }}>
+												{observation}
+											</Typography>
+
+											<Grid
+												container
+												justifyContent="space-between"
+												sx={{ mt: 2 }}>
+												<TodoCount
+													total={todos.total}
+													totalChecked={
+														todos.total_checked
+													}
+												/>
+
+												<Typography
+													variant="caption"
+													color="GrayText">
+													{createdAt}
+												</Typography>
+											</Grid>
+
+											<CardAction id="card-action">
+												<IconButton
+													onClick={() =>
+														openDialogDeleteOneNote(
+															note_id
+														)
+													}>
+													<FiTrash
+														color={
+															palette.error.main
+														}
+														size={ICON_SIZE.MEDIUM}
+													/>
+												</IconButton>
+												<Link to={linkToNoteEdit}>
+													<IconButton>
+														<FiEye
+															size={
+																ICON_SIZE.MEDIUM
+															}
+														/>
+													</IconButton>
+												</Link>
+											</CardAction>
+										</CardContent>
+									</Card>
+								</CardNoteContainer>
+							</Grid>
+						);
+					}
+				)}
+		</Grid>
 	);
 };
