+++ src/modules/notes/Note/components/CardNote/types/CardNote.types.ts
@@ -1,15 +1,6 @@
-import { Palette } from "@mui/material";
-import { INote } from "modules/notes/Note/components/CardNote/services/types";
+import { IFetchGetNotesResponse } from "modules/notes/Note/services/types";
 
-export type TypeCount = { count: number } | any;
-
-export interface ICardNotesView {
-	board_id: string | undefined;
-	palette: Palette;
-	isFetching: boolean;
-	data: {
-		count: number;
-		list_notes: INote[] | undefined;
-	};
-	openDialogDeleteOneNote: (noteID: string) => void;
+export interface ICardNotes {
+	data: IFetchGetNotesResponse | undefined;
+	isFetching?: boolean;
 }
