+++ src/modules/notes/Note/components/HeaderNote/HeaderNote.tsx
@@ -1,6 +1,5 @@
 import { useMediaQuery, useTheme } from "@mui/material";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useContextNote } from "modules/notes/Note/Context";
 import HeaderNoteView from "modules/notes/Note/components/HeaderNote/HeaderNoteView";
 import { fetchPostCreateNoteService } from "modules/notes/Note/components/HeaderNote/services";
 import { IFetchPostCreateNoteResponse } from "modules/notes/Note/components/HeaderNote/services/types";
@@ -11,14 +10,14 @@ import React, { memo } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import useFetchGetOneBoard from "shared/common/hook/useFetchGetOneBoard";
 import { createURLQueryParams } from "shared/util/createURLQueryParams";
+import { IHeaderNote } from "./types";
 
-const HeaderNote: React.FC = () => {
+const HeaderNote: React.FC<IHeaderNote> = ({ count }) => {
 	const queryClient = useQueryClient();
 	const navigate = useNavigate();
 	const { board_id } = useParams();
 
 	const { data, isFetching } = useFetchGetOneBoard(board_id);
-	const { countNotes } = useContextNote();
 
 	const { openDialogDeleteAllNotes } = useDialogNote();
 	const { breakpoints } = useTheme();
@@ -44,13 +43,15 @@ const HeaderNote: React.FC = () => {
 		{ onSuccess }
 	);
 
+	const isDisabledDeleteAllNotes = count === 0;
+
 	return (
 		<HeaderNoteView
 			{...{
 				data,
 				openDialogDeleteAllNotes,
 				isMediumScreen,
-				countNotes,
+				isDisabledDeleteAllNotes,
 				handleClickCreateNote,
 				isLoading,
 				isFetching,
