+++ src/modules/notes/Note/components/HeaderNote/types/HeaderNote.types.ts
@@ -1,12 +1,16 @@
 import { IFetchGetOneBoardResponse } from "shared/common/hook/useFetchGetOneBoard/types";
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
 
+export interface IHeaderNote {
+	count: number | undefined;
+}
+
 export interface IHeaderNoteView {
 	data: IFetchGetOneBoardResponse | undefined;
 	handleClickCreateNote: FunctionReturnsVoid;
 	openDialogDeleteAllNotes: FunctionReturnsVoid;
 	isMediumScreen: boolean;
-	countNotes: number;
+	isDisabledDeleteAllNotes: boolean;
 	isLoading: boolean;
 	isFetching: boolean;
 }
