+++ src/modules/notes/Note/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./fetchGetNotesService";
