+++ src/modules/notes/Note/services/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./fetchGetNotesService.types";
