+++ src/modules/notes/Note/types/Note.types.ts
@@ -1,0 +1,6 @@
+import { IFetchGetNotesResponse } from "../services/types";
+
+export interface INoteView {
+	data: IFetchGetNotesResponse | undefined;
+	isFetching: boolean;
+}
