--- src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/service/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./DialogDeleteAllTexts.service";
-export * from "./DialogDeleteAllTexts.service";
