+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/services/fetchDeleteAllTextsService.ts
@@ -1,0 +1,11 @@
+import { IFetchResponseDefault } from "shared/common/types/Fetch";
+import api from "shared/services/api";
+
+export const fetchDeleteAllTextsService = async (
+	boardID: string | undefined
+) => {
+	const { data } = await api.delete<IFetchResponseDefault>(
+		`/text/delete-all/board_id=${boardID}`
+	);
+	return data;
+};
