+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./fetchDeleteAllTextsService";
