+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/types/DialogDeleteAllTexts.types.ts
@@ -1,8 +1,9 @@
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
 
 export interface IDialogDeleteAllTextsView {
-	dialogDeleteAllText: { open: boolean, boardID: string };
+	dialogDeleteAllText: { open: boolean; boardID: string };
 	closeDialogDeleteAllTexts: FunctionReturnsVoid;
 	fetchDeleteAll: FunctionReturnsVoid;
-	isDeleting: boolean;
+	isLoading: boolean;
+	onClose: FunctionReturnsVoid;
 }
