+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/types/index.ts
